+++ src/node/ethereum/deposit/deposit.go
@@ -1,0 +1,57 @@
+package deposit
+
+import (
+	"context"
+	"crypto/ecdsa"
+	"github.com/BANKEX/plasma-research/src/node/ethereum/plasmacontract"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/ethclient"
+	"log"
+	"math/big"
+)
+
+func Deposit(client *ethclient.Client, privateKey string, contractAddress string, value int) string {
+	rawPrivateKey, err := crypto.HexToECDSA(privateKey)
+	if err != nil {
+		log.Println(err)
+	}
+
+	publicKey := rawPrivateKey.Public()
+	publicKeyECDSA, ok := publicKey.(*ecdsa.PublicKey)
+	if !ok {
+		log.Println("error casting public key to ECDSA")
+	}
+
+	fromAddress := crypto.PubkeyToAddress(*publicKeyECDSA)
+	nonce, err := client.PendingNonceAt(context.Background(), fromAddress)
+	if err != nil {
+		log.Println(err)
+	}
+
+	gasPrice, err := client.SuggestGasPrice(context.Background())
+	if err != nil {
+		log.Println(err)
+	}
+
+	auth := bind.NewKeyedTransactor(rawPrivateKey)
+	auth.Nonce = big.NewInt(int64(nonce))
+	auth.Value = big.NewInt(int64(value))
+	auth.GasLimit = uint64(300000)
+	auth.GasPrice = gasPrice
+
+	address := common.HexToAddress(contractAddress)
+	instance, err := store.NewStore(address, client)
+	if err != nil {
+		log.Println(err)
+	}
+
+	tx, err := instance.Deposit(auth)
+	if err != nil {
+		log.Println(err)
+	}
+
+	return tx.Hash().String()
+
+}
